methylbert/methylbert/requirements.txt
torch>=2.1
pyyaml>=6.0
orjson>=3.9.15
tqdm>=4.66
numpy>=1.23
